% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_pm.R
\name{annotate_pm}
\alias{annotate_pm}
\title{\code{annotate_pm} add Ensembl annotations to a CHIGAGO peakmatrix}
\usage{
annotate_pm(ban_file, pm_file, out_file)
}
\arguments{
\item{ban_file}{- bait annotation filename}

\item{pm_file}{- peak matrix filename}

\item{out_file}{- output filename}
}
\value{
a scalar logical depending on outcome
}
\description{
\code{annotate_pm} add Ensembl annotations to a CHIGAGO peakmatrix
}
