% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{badge_url}
\alias{badge_url}
\title{Create url for status badge}
\usage{
badge_url(status = c("concept", "experimental", "maturing", "stable"))
}
\arguments{
\item{status}{The repository status. One of "concept", "experimental",
"maturing" or "stable"}
}
\value{
url for the created badge
}
\description{
\code{badge_url} generates the URL for one of the 4 reconverse badges (concept,
experimental, maturing or stable) using shields.io.
}
\note{
Currently \code{badge_url} is hard-coded and hacky for this specific use
case but could always be generalised (although at that stage you could just
use either \href{https://github.com/matt-dray/badgr}{badgr} or
\href{https://github.com/GuangchuangYu/badger}{badger}.
}
